Start refactoring a bit
