+++ lib/chat_api/google.ex
@@ -1,53 +1,121 @@
 defmodule ChatApi.Google do
-  use OAuth2.Strategy
+  @moduledoc """
+  The Google context.
+  """
 
-  def client do
-    OAuth2.Client.new(
-      strategy: __MODULE__,
-      client_id: System.get_env("PAPERCUPS_GOOGLE_CLIENT_ID"),
-      client_secret: System.get_env("PAPERCUPS_GOOGLE_CLIENT_SECRET"),
-      redirect_uri: System.get_env("PAPERCUPS_GOOGLE_REDIRECT_URI"),
-      site: "https://accounts.google.com",
-      authorize_url: "/o/oauth2/auth",
-      token_url: "/o/oauth2/token"
-    )
+  import Ecto.Query, warn: false
+  alias ChatApi.Repo
+
+  alias ChatApi.Google.GoogleAuthorization
+
+  @doc """
+  Returns the list of google_authorizations.
+
+  ## Examples
+
+      iex> list_google_authorizations()
+      [%GoogleAuthorization{}, ...]
+
+  """
+  def list_google_authorizations do
+    Repo.all(GoogleAuthorization)
   end
 
-  def refresh_client() do
-    OAuth2.Client.new(
-      strategy: OAuth2.Strategy.Refresh,
-      client_id: System.get_env("PAPERCUPS_GOOGLE_CLIENT_ID"),
-      client_secret: System.get_env("PAPERCUPS_GOOGLE_CLIENT_SECRET"),
-      site: "https://accounts.google.com",
-      authorize_url: "/o/oauth2/auth",
-      token_url: "/o/oauth2/token"
-    )
+  @doc """
+  Gets a single google_authorization.
+
+  Raises `Ecto.NoResultsError` if the Google authorization does not exist.
+
+  ## Examples
+
+      iex> get_google_authorization!(123)
+      %GoogleAuthorization{}
+
+      iex> get_google_authorization!(456)
+      ** (Ecto.NoResultsError)
+
+  """
+  def get_google_authorization!(id), do: Repo.get!(GoogleAuthorization, id)
+
+  @doc """
+  Creates a google_authorization.
+
+  ## Examples
+
+      iex> create_google_authorization(%{field: value})
+      {:ok, %GoogleAuthorization{}}
+
+      iex> create_google_authorization(%{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def create_google_authorization(attrs \\ %{}) do
+    %GoogleAuthorization{}
+    |> GoogleAuthorization.changeset(attrs)
+    |> Repo.insert()
   end
 
-  def authorize_url!(params \\ []) do
-    OAuth2.Client.authorize_url!(client(), params)
+  @doc """
+  Updates a google_authorization.
+
+  ## Examples
+
+      iex> update_google_authorization(google_authorization, %{field: new_value})
+      {:ok, %GoogleAuthorization{}}
+
+      iex> update_google_authorization(google_authorization, %{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def update_google_authorization(%GoogleAuthorization{} = google_authorization, attrs) do
+    google_authorization
+    |> GoogleAuthorization.changeset(attrs)
+    |> Repo.update()
   end
 
-  # You can pass options to the underlying http library via `opts` parameter
-  def get_token!(params \\ [], headers \\ [], opts \\ []) do
-    case params do
-      [refresh_token: refresh_token] when not is_nil(refresh_token) ->
-        OAuth2.Client.get_token!(refresh_client(), params, headers, opts)
+  @doc """
+  Deletes a google_authorization.
 
-      _ ->
-        OAuth2.Client.get_token!(client(), params, headers, opts)
+  ## Examples
+
+      iex> delete_google_authorization(google_authorization)
+      {:ok, %GoogleAuthorization{}}
+
+      iex> delete_google_authorization(google_authorization)
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_google_authorization(%GoogleAuthorization{} = google_authorization) do
+    Repo.delete(google_authorization)
+  end
+
+  @doc """
+  Returns an `%Ecto.Changeset{}` for tracking google_authorization changes.
+
+  ## Examples
+
+      iex> change_google_authorization(google_authorization)
+      %Ecto.Changeset{data: %GoogleAuthorization{}}
+
+  """
+  def change_google_authorization(%GoogleAuthorization{} = google_authorization, attrs \\ %{}) do
+    GoogleAuthorization.changeset(google_authorization, attrs)
+  end
+
+  def get_authorization_by_account(account_id) do
+    GoogleAuthorization
+    |> where(account_id: ^account_id)
+    |> order_by(desc: :inserted_at)
+    |> Repo.one()
+  end
+
+  def create_or_update_authorization(account_id, params) do
+    existing = get_authorization_by_account(account_id)
+
+    if existing do
+      update_google_authorization(existing, params)
+    else
+      create_google_authorization(params)
     end
-  end
-
-  # Strategy Callbacks
-  def authorize_url(client, params) do
-    OAuth2.Strategy.AuthCode.authorize_url(client, params)
-  end
-
-  def get_token(client, params \\ [], headers \\ []) do
-    client
-    |> put_param(:client_secret, client.client_secret)
-    |> put_header("accept", "application/json")
-    |> OAuth2.Strategy.AuthCode.get_token(params, headers)
   end
 end
